Code health (4/5): python behavior fixes — desloppify sweep
Splits the python/ slice of the desloppify-integration consolidation (#474,
post-integration) into two stacked branches. This is the BEHAVIOR branch.

Contents (all under python/):
- agent/ extraction, lifecycle, debrief, state, loop_runner, _text,
  contradiction behavior edits
- client.py, operations.py (WalletContext introduced — internal to the
  client↔operations layer), relay.py, reranker.py (PyStemmer dropped),
  claims_helper.py, retype_setscope.py, memory_types.py, pair/* helpers
- crystals/ package move (recrystallize + session_segmentation) with the
  3 runtime imports.engine references rewritten back to import_engine
  (the unsplit location); root shims alias the new package
- hermes/tools.py reconstructed UNSPLIT: origin/main's header + import
  subsystem, with only the non-import handlers (remember, recall, forget,
  status, setup, upgrade, top_up) updated to the sweep's accessor/ForgetResult
  behavior. No reference to hermes.import_tools.
- behavior test changes + new tests (test_extraction_helpers, test_pair_page)

Deliberately EXCLUDED (they land in 5/5, split/pr-e):
- imports/ package, shim versions of import_engine/import_state/_smart_import/
  import_adapters/__init__, hermes/import_tools.py, the split tools.py,
  test_tools_helpers.py, and the desloppify versions of the
  structural-referencing adapter/smart-import tests.

The import subsystem (import_engine.py, _smart_import.py, import_state.py,
import_adapters/) stays at origin/main here; the structural-referencing
tests stay at their origin/main versions to match. client.remember() — the
bridge the import engine calls — is signature-identical main↔desloppify, so
the unsplit import subsystem is consistent with the new client internals.

Verified: full python suite green (2075 passed, 39 skipped, 1 xfailed);
`grep -r totalreclaw.imports|hermes.import_tools|chunk_skip_reason python/`
returns nothing.

Stacked under: split/pr-d ← origin/main; 5/5 (split/pr-e) applies the
import-subsystem restructure on top of this.
