+++ python/pyproject.toml
@@ -56,7 +56,6 @@ dependencies = [
     "httpx>=0.27",
     "onnxruntime>=1.24",
     "numpy>=1.26",
-    "PyStemmer>=2.2",
     "transformers>=4.40",
     "huggingface-hub>=0.23",
     "eth-account>=0.13",
